+++ src/main/java/io/github/hananjafari76/sample/category/repository/CategoryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    Optional<Category> findById(int id);
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    Optional<Category> findById(Long id);
 }
